requirements.txt
# Flask API Dependencies
Flask==3.0.0
Flask-CORS==4.0.0
joblib==1.3.2
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
imbalanced-learn==0.11.0
gunicorn==21.2.0

# MySQL connector (optional - if you connect to DB from Flask)
mysql-connector-python==8.2.0
